Parse listening stats durations as fractional seconds
totalTime, dayOfWeek, item and session times arrive as fractional seconds and
`days` maps a date straight to seconds. Declared as Long and DayStats, the
response never parsed, so the stats screens never loaded.

+++ app/src/main/java/com/paulohenriquesg/fahrenheit/main/MainScreen.kt
@@ -845,7 +845,7 @@ fun StatsBrowseView(stats: com.paulohenriquesg.fahrenheit.api.ListeningStatsResp
             ) {
                 com.paulohenriquesg.fahrenheit.stats.StatCard(
                     title = "Total Listening Time",
-                    value = com.paulohenriquesg.fahrenheit.stats.formatTime(stats.totalTime)
+                    value = com.paulohenriquesg.fahrenheit.stats.formatTime(stats.totalTime.toLong())
                 )
 
                 stats.items.size.let { itemCount ->
@@ -863,7 +863,7 @@ fun StatsBrowseView(stats: com.paulohenriquesg.fahrenheit.api.ListeningStatsResp
                 }
 
                 if (stats.days.isNotEmpty()) {
-                    val avgPerDay = stats.totalTime / stats.days.size
+                    val avgPerDay = (stats.totalTime / stats.days.size).toLong()
                     com.paulohenriquesg.fahrenheit.stats.StatCard(
                         title = "Average per Day",
                         value = com.paulohenriquesg.fahrenheit.stats.formatTime(avgPerDay)
